Enforce the sign-in guard on the user recipe list
